+++ db-setup.sql
@@ -134,3 +134,32 @@ ALTER TABLE `kistaai_users__google`
 
 ALTER TABLE `kistaai_users__sessions`
 	ADD COLUMN `user_agent` VARCHAR(255) NOT NULL DEFAULT '' AFTER `valid_to`;
+
+/* * * * * * */
+
+CREATE TABLE `kistaai_replicate__uploads` (
+	`upload_id` INT(10) UNSIGNED NOT NULL AUTO_INCREMENT,
+	`user_id` INT(10) NULL DEFAULT NULL,
+	`created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
+	`updated` DATETIME NULL DEFAULT NULL,
+	`realname` VARCHAR(255) NULL DEFAULT NULL COMMENT 'Presentation' COLLATE 'utf8mb4_danish_ci',
+	`filename` VARCHAR(255) NULL DEFAULT NULL COMMENT 'On disk' COLLATE 'utf8mb4_danish_ci',
+	`extension` VARCHAR(32) NULL DEFAULT NULL COLLATE 'utf8mb4_danish_ci',
+	`filesize` INT(10) UNSIGNED NOT NULL DEFAULT '0',
+	`thumbnail` VARCHAR(255) NOT NULL DEFAULT '' COLLATE 'utf8mb4_danish_ci',
+	`status` VARCHAR(64) NOT NULL DEFAULT '' COLLATE 'utf8mb4_danish_ci',
+	PRIMARY KEY (`upload_id`) USING BTREE
+) COLLATE='utf8mb4_danish_ci' ENGINE=InnoDB AUTO_INCREMENT=1;
+
+CREATE TABLE `kistaai_replicate__images` (
+	`image_id` INT(10) UNSIGNED NOT NULL AUTO_INCREMENT,
+	`upload_id` INT(10) NULL DEFAULT NULL,
+	`user_id` INT(10) NULL DEFAULT NULL,
+	`created` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
+	`filename` VARCHAR(255) NULL DEFAULT NULL COMMENT 'On disk' COLLATE 'utf8mb4_danish_ci',
+	`extension` VARCHAR(32) NULL DEFAULT NULL COLLATE 'utf8mb4_danish_ci',
+	`filesize` INT(10) UNSIGNED NOT NULL DEFAULT '0',
+	`thumbnail` VARCHAR(255) NOT NULL DEFAULT '' COLLATE 'utf8mb4_danish_ci',
+	`status` VARCHAR(64) NOT NULL DEFAULT '' COLLATE 'utf8mb4_danish_ci',
+	PRIMARY KEY (`image_id`) USING BTREE
+) COLLATE='utf8mb4_danish_ci' ENGINE=InnoDB AUTO_INCREMENT=1;