+++ htdocs/inc_appdata/openai_helpers.php
@@ -424,3 +424,6 @@ function promptChatGPT4($prompt){
     return promptChatGPT($prompt, 'gpt-4-1106-preview');
 }
 
+
+
+