+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Added, Changed, Deprecated, Removed, Fixed, Securit
 ### v0.10.0
 - added, extra webfolder for NGROK tunnel and Replicate webhooks for development
 - added, POC for inference.
+- added, new upload feature for style transfer / avatar mode.
 
 ### v0.9.6
 - fixed, Moving action sheets, snackbars and sidebars outside page-content